+++ CHANGELOG.md
@@ -22,6 +22,7 @@ to the [Haskell Package Versioning Policy][pvp].
   So that you can send synchronous/blocking messages that don't have a
   meaningful response. Useful if you need to wait until a message is handled
   before proceeding.
+- Loosened dependency version bounds
 
 ## [0.3.0.0] - 2021-02-23