+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog][changelog], and this project adheres
 to the [Haskell Package Versioning Policy][pvp].
 
-## [0.4.0.0] - Unreleased
+## [0.4.0.0] - 2021-05-19
 
 - Remove low-level `Process` API in favor of high-level `Server` API
 - Rename "process" to "actor" everywhere
@@ -102,7 +102,8 @@ don't receive messages
 
 Initial release
 
-[0.4.0.0]: https://github.com/evanrelf/drama/compare/v0.3.0.0...HEAD
+[Unreleased]: [https://github.com/evanrelf/drama/compare/v0.4.0.0...HEAD]
+[0.4.0.0]: https://github.com/evanrelf/drama/releases/tag/v0.4.0.0
 [0.3.0.0]: https://github.com/evanrelf/drama/releases/tag/v0.3.0.0
 [0.2.0.0]: https://github.com/evanrelf/drama/releases/tag/v0.2.0.0
 [0.1.0.3]: https://github.com/evanrelf/drama/releases/tag/v0.1.0.3